+++ dto/file.go
@@ -10,7 +10,7 @@ type FileDTO struct {
 	ParentID    *int    `json:"parent_id"`
 	Name        string  `json:"name"`
 	Description *string `json:"description"`
-	Size        int     `json:"size"`
+	Size        int64   `json:"size"`
 	Type        string  `json:"type"`
 }
 
@@ -19,7 +19,7 @@ type FileResponseDTO struct {
 	ParentID    *int      `json:"parent_id"`
 	Name        string    `json:"name"`
 	Description *string   `json:"description"`
-	Size        int       `json:"size"`
+	Size        int64     `json:"size"`
 	Type        string    `json:"type"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
